+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.0
+- New configuration design
+- New versions of jQuery and jQuery UI
+- Set the position of the digits with drag & drop
+- Configure for the clock two gradients, too.
+- Set all colors with a color picker
+- Refactoring, optimization
+
 ## 1.0
 - Counter
 - Clock